refactor naming convention
