+++ docs/src/stories/components/Layout/LayoutBeta.stories.jsx
@@ -1,18 +1,16 @@
 import React from 'react'
 import clsx from 'clsx'
-import {NavWithSubItems} from '../ActionList/ActionListPatterns.stories'
 import PageLayoutBehavior from '../../helpers/pageLayoutBehavior.jsx'
 
 export default {
   title: 'Components/Layout/Beta',
   excludeStories: ['LayoutTemplate'],
   argTypes: {
-
     // Debug
 
     _debug: {
       control: 'boolean',
-      description: 'Show background colors in regions for debugging',
+      description: 'Show background colors in regions for debugging'
     },
 
     // Structure
@@ -23,7 +21,8 @@ export default {
         type: 'inline-radio',
         labels: ['fluid', 'md', 'lg', 'xl']
       },
-      description: 'Define the maximum width of the component. `fluid` sets it to full-width. Other values center `Layout` horizontally. Refer to [container utilities](https://primer.style/css/objects/grid#containers) for reference.',
+      description:
+        'Define the maximum width of the component. `fluid` sets it to full-width. Other values center `Layout` horizontally. Refer to [container utilities](https://primer.style/css/objects/grid#containers) for reference.',
       table: {
         category: 'Structure'
       }
@@ -33,7 +32,8 @@ export default {
       control: {
         type: 'inline-radio'
       },
-      description: 'Sets wrapper margins surrounding the component to distance itself from the viewport edges. `normal` sets the margin to 16px, and to 24px on `lg` breakpoints and above. `condensed` keeps the margin at 16px. `none` sets the margin to 0.',
+      description:
+        'Sets wrapper margins surrounding the component to distance itself from the viewport edges. `normal` sets the margin to 16px, and to 24px on `lg` breakpoints and above. `condensed` keeps the margin at 16px. `none` sets the margin to 0.',
       table: {
         category: 'Structure'
       }
@@ -43,7 +43,8 @@ export default {
       control: {
         type: 'inline-radio'
       },
-      description: 'Sets padding to regions individually. `normal` sets padding to 16px, with the `content` region getting 24px horizontal padding on `lg` breakpoints and above. `condensed` keeps the padding always at `16px`. `none` sets the padding to 0.',
+      description:
+        'Sets padding to regions individually. `normal` sets padding to 16px, with the `content` region getting 24px horizontal padding on `lg` breakpoints and above. `condensed` keeps the padding always at `16px`. `none` sets the padding to 0.',
       table: {
         category: 'Structure'
       }
@@ -53,7 +54,8 @@ export default {
       control: {
         type: 'inline-radio'
       },
-      description: 'Sets the gap between columns to distance them from each other. `normal` sets the gap to 16px, and to 24px on `lg` breakpoints and above. `condensed` keeps the gap always at 16px. `none` sets the gap to 0.',
+      description:
+        'Sets the gap between columns to distance them from each other. `normal` sets the gap to 16px, and to 24px on `lg` breakpoints and above. `condensed` keeps the gap always at 16px. `none` sets the gap to 0.',
       table: {
         category: 'Structure'
       }
@@ -63,7 +65,8 @@ export default {
       control: {
         type: 'inline-radio'
       },
-      description: 'Sets the gap below the header and above the footer. `normal` sets the gap to 16px, and to 24px on `lg` breakpoints and above. `condensed` keeps the gap always at 16px. `none` sets the gap to 0.',
+      description:
+        'Sets the gap below the header and above the footer. `normal` sets the gap to 16px, and to 24px on `lg` breakpoints and above. `condensed` keeps the gap always at 16px. `none` sets the gap to 0.',
       table: {
         category: 'Structure'
       }
@@ -76,7 +79,8 @@ export default {
       control: {
         type: 'inline-radio'
       },
-      description: '`responsiveVariant` defines how the layout component adapts to smaller viewports. `stackRegions` presents the content in a vertical flow, with `pane` and `content` vertically arranged. `separateRegions` presents `pane` and `content` as different pages on smaller viewports. Change the preview size from the toolbar to test it.',
+      description:
+        '`responsiveVariant` defines how the layout component adapts to smaller viewports. `stackRegions` presents the content in a vertical flow, with `pane` and `content` vertically arranged. `separateRegions` presents `pane` and `content` as different pages on smaller viewports. Change the preview size from the toolbar to test it.',
       table: {
         category: 'Responsive'
       }
@@ -86,7 +90,8 @@ export default {
       control: {
         type: 'inline-radio'
       },
-      description: 'When `responsiveVariant` is set to `separateRegions`, defines which region appears first on small viewports. `content` is default.',
+      description:
+        'When `responsiveVariant` is set to `separateRegions`, defines which region appears first on small viewports. `content` is default.',
       table: {
         category: 'Responsive'
       }
@@ -97,7 +102,7 @@ export default {
     paneWidth: {
       options: ['default', 'narrow', 'wide'],
       control: {
-      type: 'inline-radio'
+        type: 'inline-radio'
       },
       description: 'Defines the width of the pane',
       table: {
@@ -109,23 +114,25 @@ export default {
       control: {
         type: 'inline-radio'
       },
-      description: 'Defines the position of the pane. `start` renders the pane on the left, and `end` renders it on the right.',
+      description:
+        'Defines the position of the pane. `start` renders the pane on the left, and `end` renders it on the right.',
       table: {
-        category: 'Pane',
+        category: 'Pane'
       }
     },
     panePositionNarrow: {
       options: ['inherit', 'start', 'end'],
       control: {
-        type: 'inline-radio',
+        type: 'inline-radio'
       },
-      description: 'If `responsiveVariant` is set to `stackRegions`, defines the position of the pane in narrow viewports. `start` puts the pane above `content`, and `end` puts it below `content`. `inherit` uses the same value from `panePosition`.',
+      description:
+        'If `responsiveVariant` is set to `stackRegions`, defines the position of the pane in narrow viewports. `start` puts the pane above `content`, and `end` puts it below `content`. `inherit` uses the same value from `panePosition`.',
       table: {
         category: 'Pane'
       }
     },
     hasPaneDivider: {
-      control: { type: 'boolean' },
+      control: {type: 'boolean'},
       description: 'Whether to show a pane line divider.',
       table: {
         category: 'Pane'
@@ -136,13 +143,14 @@ export default {
       control: {
         type: 'inline-radio'
       },
-      description: 'Whether to show a divider between `pane` and `content` regions if `responsiveVariant` is set to `stackRegions`. `line` shows a single line. `filled` shows a thicker mobile-frienldy divider.',
+      description:
+        'Whether to show a divider between `pane` and `content` regions if `responsiveVariant` is set to `stackRegions`. `line` shows a single line. `filled` shows a thicker mobile-frienldy divider.',
       table: {
         category: 'Pane'
       }
     },
     paneIsSticky: {
-      control: { type: 'boolean' },
+      control: {type: 'boolean'},
       description: 'Whether to make the pane sticky.',
       table: {
         category: 'Pane'
@@ -156,24 +164,24 @@ export default {
       control: {
         type: 'inline-radio'
       },
-      description: 'Defines the maximum width of the content region. `fluid` sets it to full-width. Other values follow container widths from `sm` to `xl`. With smaller widths, the content region will try to stay centered to the viewport area.',
+      description:
+        'Defines the maximum width of the content region. `fluid` sets it to full-width. Other values follow container widths from `sm` to `xl`. With smaller widths, the content region will try to stay centered to the viewport area.',
       table: {
         category: 'Content'
       }
     },
 
-
     // Header
 
     hasHeader: {
-      control: { type: 'boolean' },
+      control: {type: 'boolean'},
       table: {
         category: 'Header'
       }
     },
 
     hasHeaderDivider: {
-      control: { type: 'boolean' },
+      control: {type: 'boolean'},
       description: 'Whether to show a header divider.',
       table: {
         category: 'Header'
@@ -181,28 +189,28 @@ export default {
     },
 
     headerDividerNarrow: {
-    options: ['inherit', 'none', 'line', 'filled'],
+      options: ['inherit', 'none', 'line', 'filled'],
       control: {
         type: 'inline-radio'
       },
-      description: 'Defines how the `header` divider should look on narrow viewports. `inherit` renders a `line` if `hasHeaderDivider` is true. `filled` shows a thicker mobile-friendly divider.',
+      description:
+        'Defines how the `header` divider should look on narrow viewports. `inherit` renders a `line` if `hasHeaderDivider` is true. `filled` shows a thicker mobile-friendly divider.',
       table: {
         category: 'Header'
       }
     },
 
-
     // Footer
 
     hasFooter: {
-      control: { type: 'boolean' },
+      control: {type: 'boolean'},
       table: {
         category: 'Footer'
       }
     },
 
     hasFooterDivider: {
-      control: { type: 'boolean' },
+      control: {type: 'boolean'},
       description: 'Whether to show a footer divider.',
       table: {
         category: 'Footer'
@@ -214,7 +222,8 @@ export default {
       control: {
         type: 'inline-radio'
       },
-      description: 'Whether to show a divider above the `footer` region on narrow viewports. `line` shows a single line. `filled` shows a thicker mobile-frienldy divider.',
+      description:
+        'Whether to show a divider above the `footer` region on narrow viewports. `line` shows a single line. `filled` shows a thicker mobile-frienldy divider.',
       table: {
         category: 'Footer'
       }
@@ -249,14 +258,13 @@ export default {
   }
 }
 
-const layoutClassName = 'PageLayout';
+const layoutClassName = 'PageLayout'
 
 // build every component case here in the template (private api)
 export const LayoutTemplate = ({
-
   // Debug
   _debug,
-  
+
   // Wrapper
   wrapperSizing,
 
@@ -265,7 +273,7 @@ export const LayoutTemplate = ({
   innerSpacing,
   columnGap,
   rowGap,
-  
+
   // Pane
   paneWidth,
   panePosition,
@@ -297,152 +305,160 @@ export const LayoutTemplate = ({
   paneChildren,
   footerChildren
 }) => {
-
   const containerClass = {
-    'full': '',
-    'md': 'container-md',
-    'lg': 'container-lg',
-    'xl': 'container-xl'
-  };
+    full: '',
+    md: 'container-md',
+    lg: 'container-lg',
+    xl: 'container-xl'
+  }
 
   // Default values
-  wrapperSizing = wrapperSizing ?? 'xl';
-  outerSpacing = outerSpacing ?? 'normal';
-  innerSpacing = innerSpacing ?? 'none';
-  columnGap = columnGap ?? 'normal';
-  rowGap = rowGap ?? 'normal';
-  panePosition = panePosition ?? 'end';
-  panePositionNarrow = panePositionNarrow ?? 'inherit';
-  responsiveVariant = responsiveVariant ?? 'stackRegions';
-  primaryRegion = primaryRegion ?? 'content';
+  wrapperSizing = wrapperSizing ?? 'xl'
+  outerSpacing = outerSpacing ?? 'normal'
+  innerSpacing = innerSpacing ?? 'none'
+  columnGap = columnGap ?? 'normal'
+  rowGap = rowGap ?? 'normal'
+  panePosition = panePosition ?? 'end'
+  panePositionNarrow = panePositionNarrow ?? 'inherit'
+  responsiveVariant = responsiveVariant ?? 'stackRegions'
+  primaryRegion = primaryRegion ?? 'content'
 
   // Leave `null` values for states that don't require a modifier class
-  outerSpacing = (outerSpacing === 'none') ? null : outerSpacing;
-  innerSpacing = (innerSpacing === 'none') ? null : innerSpacing;
-  paneWidth = (paneWidth === 'default') ? null : paneWidth;
-  contentWidth = (contentWidth === 'fluid') ? null : contentWidth;
-  headerDividerNarrow = (headerDividerNarrow === 'none') ? null : headerDividerNarrow;
-  footerDividerNarrow = (footerDividerNarrow === 'none') ? null : footerDividerNarrow;
+  outerSpacing = outerSpacing === 'none' ? null : outerSpacing
+  innerSpacing = innerSpacing === 'none' ? null : innerSpacing
+  paneWidth = paneWidth === 'default' ? null : paneWidth
+  contentWidth = contentWidth === 'fluid' ? null : contentWidth
+  headerDividerNarrow = headerDividerNarrow === 'none' ? null : headerDividerNarrow
+  footerDividerNarrow = footerDividerNarrow === 'none' ? null : footerDividerNarrow
 
   // Inherit value for responsive props
-  panePositionNarrow = (panePositionNarrow === 'inherit') ? panePosition : panePositionNarrow;
+  panePositionNarrow = panePositionNarrow === 'inherit' ? panePosition : panePositionNarrow
 
   if (hasPaneDivider) {
-    paneDividerNarrow = (paneDividerNarrow === 'inherit') ? 'line' : paneDividerNarrow;
+    paneDividerNarrow = paneDividerNarrow === 'inherit' ? 'line' : paneDividerNarrow
   } else {
-    paneDividerNarrow = null;
+    paneDividerNarrow = null
   }
 
   if (hasHeaderDivider) {
-    headerDividerNarrow = (headerDividerNarrow === 'inherit') ? 'line' : headerDividerNarrow;
+    headerDividerNarrow = headerDividerNarrow === 'inherit' ? 'line' : headerDividerNarrow
   } else {
-    headerDividerNarrow = null;
+    headerDividerNarrow = null
   }
 
   if (hasFooterDivider) {
-    footerDividerNarrow = (footerDividerNarrow === 'inherit') ? 'line' : footerDividerNarrow;
+    footerDividerNarrow = footerDividerNarrow === 'inherit' ? 'line' : footerDividerNarrow
   } else {
-    footerDividerNarrow = null;
+    footerDividerNarrow = null
   }
 
-  PageLayoutBehavior();
+  PageLayoutBehavior()
 
   return (
     <>
       <div
-      className={clsx(
-        layoutClassName,
-        
-        outerSpacing && layoutClassName + '--outerSpacing-' + `${outerSpacing}`,
-        innerSpacing && layoutClassName + '--innerSpacing-' + `${innerSpacing}`,
-        columnGap && layoutClassName + '--columnGap-' + `${columnGap}`,
-        rowGap && layoutClassName + '--rowGap-' + `${rowGap}`,
+        className={clsx(
+          layoutClassName,
 
-        paneWidth && layoutClassName + '--paneWidth-' + `${paneWidth}`,
-        panePosition && layoutClassName + '--panePos-' + `${panePosition}`,
-        hasPaneDivider && layoutClassName + '--hasPaneDivider',
-        paneIsSticky && layoutClassName + '--isPaneSticky',
+          outerSpacing && layoutClassName + '--outerSpacing-' + `${outerSpacing}`,
+          innerSpacing && layoutClassName + '--innerSpacing-' + `${innerSpacing}`,
+          columnGap && layoutClassName + '--columnGap-' + `${columnGap}`,
+          rowGap && layoutClassName + '--rowGap-' + `${rowGap}`,
 
-        layoutClassName + '--responsive-' + `${responsiveVariant}`,
-        responsiveVariant === 'separateRegions' && layoutClassName + '--responsive-primary-' + `${primaryRegion}`,
-        responsiveVariant === 'stackRegions' && panePositionNarrow && layoutClassName + '--responsive-panePos-' + `${panePositionNarrow}`,
-      )}
+          paneWidth && layoutClassName + '--paneWidth-' + `${paneWidth}`,
+          panePosition && layoutClassName + '--panePos-' + `${panePosition}`,
+          hasPaneDivider && layoutClassName + '--hasPaneDivider',
+          paneIsSticky && layoutClassName + '--isPaneSticky',
+
+          layoutClassName + '--responsive-' + `${responsiveVariant}`,
+          responsiveVariant === 'separateRegions' && layoutClassName + '--responsive-primary-' + `${primaryRegion}`,
+          responsiveVariant === 'stackRegions' &&
+            panePositionNarrow &&
+            layoutClassName + '--responsive-panePos-' + `${panePositionNarrow}`
+        )}
       >
-        <div className={clsx(
-          layoutClassName + '-wrapper',
-          wrapperSizing && containerClass[wrapperSizing]
-          )}>
-            
+        <div className={clsx(layoutClassName + '-wrapper', wrapperSizing && containerClass[wrapperSizing])}>
           {/* Header */}
-          {hasHeader &&
-            <div className={clsx(
-              layoutClassName + '-region',
-              layoutClassName + '-header',
-              hasHeaderDivider && layoutClassName + '-header--hasDivider',
-              headerDividerNarrow && layoutClassName + '-region--dividerNarrow-' + headerDividerNarrow + '-after'
-            )}>
+          {hasHeader && (
+            <div
+              className={clsx(
+                layoutClassName + '-region',
+                layoutClassName + '-header',
+                hasHeaderDivider && layoutClassName + '-header--hasDivider',
+                headerDividerNarrow && layoutClassName + '-region--dividerNarrow-' + headerDividerNarrow + '-after'
+              )}
+            >
               {headerChildren}
             </div>
-          }
+          )}
 
-          <div className={clsx(
-            layoutClassName + '-columns'
-          )}>
-
+          <div className={clsx(layoutClassName + '-columns')}>
             {/* pane if rendered first */}
-            {panePosition === 'start' &&
-              <div className={clsx(
-                layoutClassName + '-region',
-                layoutClassName + '-pane',
-                paneDividerNarrow && layoutClassName + '-region--dividerNarrow-' + paneDividerNarrow + (panePositionNarrow === 'start' ? '-after' : '-before') 
-              )}>
+            {panePosition === 'start' && (
+              <div
+                className={clsx(
+                  layoutClassName + '-region',
+                  layoutClassName + '-pane',
+                  paneDividerNarrow &&
+                    layoutClassName +
+                      '-region--dividerNarrow-' +
+                      paneDividerNarrow +
+                      (panePositionNarrow === 'start' ? '-after' : '-before')
+                )}
+              >
                 {paneChildren}
               </div>
-            }
+            )}
 
             {/* content */}
-            <div className={clsx(
-              layoutClassName + '-region',
-              layoutClassName + '-content'
-            )}>
+            <div className={clsx(layoutClassName + '-region', layoutClassName + '-content')}>
               {contentWidth ? (
-              <>
-                <div className={layoutClassName + '-content-centered-' + contentWidth}>
-                  <div className={'container-' + contentWidth}>
-                    {contentChildren}
+                <>
+                  <div className={layoutClassName + '-content-centered-' + contentWidth}>
+                    <div className={'container-' + contentWidth}>{contentChildren}</div>
                   </div>
-                </div>
-              </>
+                </>
               ) : (
-              <>
-                {contentChildren}
-              </>
+                <>{contentChildren}</>
               )}
             </div>
 
             {/* pane if rendered last */}
-            {panePosition === 'end' &&
-            <div className={clsx(
-              layoutClassName + '-region',
-              layoutClassName + '-pane',
-              paneDividerNarrow && layoutClassName + '-region--dividerNarrow-' + paneDividerNarrow + (panePositionNarrow === 'start' ? '-after' : '-before')
-            )}>
-              {paneChildren}
-            </div>}
+            {panePosition === 'end' && (
+              <div
+                className={clsx(
+                  layoutClassName + '-region',
+                  layoutClassName + '-pane',
+                  paneDividerNarrow &&
+                    layoutClassName +
+                      '-region--dividerNarrow-' +
+                      paneDividerNarrow +
+                      (panePositionNarrow === 'start' ? '-after' : '-before')
+                )}
+              >
+                {paneChildren}
+              </div>
+            )}
           </div>
 
           {/* footer */}
-          {hasFooter && <div className={clsx(
-            layoutClassName + '-region',
-            layoutClassName + '-footer',
-            hasFooterDivider && layoutClassName + '-footer--hasDivider',
-            footerDividerNarrow && layoutClassName + '-region--dividerNarrow-' + footerDividerNarrow + '-before'
-            )}>{footerChildren}</div>}
-          </div>
+          {hasFooter && (
+            <div
+              className={clsx(
+                layoutClassName + '-region',
+                layoutClassName + '-footer',
+                hasFooterDivider && layoutClassName + '-footer--hasDivider',
+                footerDividerNarrow && layoutClassName + '-region--dividerNarrow-' + footerDividerNarrow + '-before'
+              )}
+            >
+              {footerChildren}
+            </div>
+          )}
+        </div>
 
-          {/* debug */}
-          {_debug &&
-            <style type='text/css'>{`
+        {/* debug */}
+        {_debug && (
+          <style type="text/css">{`
               .PageLayout-header {
                 background: lightpink;
               }
@@ -456,17 +472,17 @@ export const LayoutTemplate = ({
                 background: lightyellow;
               }
             `}</style>
-          }
+        )}
       </div>
     </>
-  );
-};
+  )
+}
 
-export const Playground = LayoutTemplate.bind({});
-Playground.storyName = 'Playground';
+export const Playground = LayoutTemplate.bind({})
+Playground.storyName = 'Playground'
 Playground.parameters = {
-  layout: 'fullscreen',
-};
+  layout: 'fullscreen'
+}
 Playground.args = {
   _debug: true,
   wrapperSizing: 'xl',
